Add very simple UI for installing .NET and launching UEVR
There are a lot of missing components here. The details page can't
be closed (and could be prettified), I still haven't implemented
UEVR installation management, launching any executable seems to
fail if the game hasn't been launched once with the current Proton
version, etc. But it does show signs of working, which is a good thing.